+++ src/lib.rs
@@ -12,16 +12,15 @@ pub mod prelude;
 pub mod tools;
 pub mod utils;
 
-
 #[cfg(feature = "python")]
 mod python;
-
 
 lazy_static! {
     static ref LOG: u64 = init_log();
 }
 
 // install a logger facility
+#[allow(unused)]
 fn init_log() -> u64 {
     let _res = env_logger::try_init();
     log::info!("\n ************** initializing logger *****************\n");