+++ Changes.md
@@ -1,3 +1,7 @@
+- version 0.1.6  
+    -- The embedding computes a continuity ratio for each point  
+    -- The plotCsvLabels now dumps a map of continuity ratio enabling quality inspection in a graphical way (see visu.jl)
+
 - version 0.1.5
   
     -- switched to edition 2024